Group ids with peer reviews should not be deleted

+++ orcid-core/src/main/java/org/orcid/core/manager/impl/GroupIdRecordManagerImpl.java
@@ -22,6 +22,7 @@ import javax.annotation.Resource;
 
 import org.orcid.core.exception.DuplicatedGroupIdRecordException;
 import org.orcid.core.exception.GroupIdRecordNotFoundException;
+import org.orcid.core.exception.OrcidElementCantBeDeletedException;
 import org.orcid.core.locale.LocaleManager;
 import org.orcid.core.manager.GroupIdRecordManager;
 import org.orcid.core.manager.OrcidSecurityManager;
@@ -45,15 +46,15 @@ public class GroupIdRecordManagerImpl extends GroupIdRecordManagerReadOnlyImpl i
 
     @Resource
     private OrcidSecurityManager orcidSecurityManager;
-    
-    @Resource 
+
+    @Resource
     private ActivityValidator activityValidator;
 
     @Override
     public GroupIdRecord createGroupIdRecord(GroupIdRecord groupIdRecord) {
-    	SourceEntity sourceEntity = sourceManager.retrieveSourceEntity();
-    	activityValidator.validateGroupIdRecord(groupIdRecord, true, sourceEntity);
-    	validateDuplicate(groupIdRecord);    	
+        SourceEntity sourceEntity = sourceManager.retrieveSourceEntity();
+        activityValidator.validateGroupIdRecord(groupIdRecord, true, sourceEntity);
+        validateDuplicate(groupIdRecord);
         if (sourceEntity != null) {
             Source source = new Source();
             if (sourceEntity.getSourceClient() != null) {
@@ -63,7 +64,7 @@ public class GroupIdRecordManagerImpl extends GroupIdRecordManagerReadOnlyImpl i
             }
             groupIdRecord.setSource(source);
         }
-        
+
         GroupIdRecordEntity entity = jpaJaxbGroupIdRecordAdapter.toGroupIdRecordEntity(groupIdRecord);
         groupIdRecordDao.persist(entity);
         return jpaJaxbGroupIdRecordAdapter.toGroupIdRecord(entity);
@@ -71,27 +72,27 @@ public class GroupIdRecordManagerImpl extends GroupIdRecordManagerReadOnlyImpl i
 
     @Override
     public GroupIdRecord updateGroupIdRecord(Long putCode, GroupIdRecord groupIdRecord) {
-        GroupIdRecordEntity existingEntity = groupIdRecordDao.find(putCode);  
-        
-        if(existingEntity == null) {
+        GroupIdRecordEntity existingEntity = groupIdRecordDao.find(putCode);
+
+        if (existingEntity == null) {
             throw new GroupIdRecordNotFoundException();
-        }  
-        
+        }
+
         SourceEntity sourceEntity = sourceManager.retrieveSourceEntity();
-        //Save the original source
+        // Save the original source
         String existingSourceId = existingEntity.getSourceId();
-        String existingClientSourceId = existingEntity.getClientSourceId();             
-        
-        activityValidator.validateGroupIdRecord(groupIdRecord, false, sourceEntity);        
+        String existingClientSourceId = existingEntity.getClientSourceId();
+
+        activityValidator.validateGroupIdRecord(groupIdRecord, false, sourceEntity);
         validateDuplicate(groupIdRecord);
-        
+
         orcidSecurityManager.checkSource(existingEntity);
         GroupIdRecordEntity updatedEntity = jpaJaxbGroupIdRecordAdapter.toGroupIdRecordEntity(groupIdRecord);
         updatedEntity.setDateCreated(existingEntity.getDateCreated());
-        //Be sure it doesn't overwrite the source
+        // Be sure it doesn't overwrite the source
         updatedEntity.setSourceId(existingSourceId);
         updatedEntity.setClientSourceId(existingClientSourceId);
-        
+
         updatedEntity = groupIdRecordDao.merge(updatedEntity);
         return jpaJaxbGroupIdRecordAdapter.toGroupIdRecord(updatedEntity);
     }
@@ -100,23 +101,27 @@ public class GroupIdRecordManagerImpl extends GroupIdRecordManagerReadOnlyImpl i
     public void deleteGroupIdRecord(Long putCode) {
         GroupIdRecordEntity existingEntity = groupIdRecordDao.find(putCode);
         if (existingEntity != null) {
+            if (groupIdRecordDao.haveAnyPeerReview(existingEntity.getGroupId())) {
+                throw new OrcidElementCantBeDeletedException("Unable to delete group id because there are peer reviews associated to it");
+            }
             orcidSecurityManager.checkSource(existingEntity);
             groupIdRecordDao.remove(Long.valueOf(putCode));
         } else {
             throw new GroupIdRecordNotFoundException();
         }
-    }    
+    }
 
     private void validateDuplicate(GroupIdRecord newGroupIdRecord) {
         List<GroupIdRecordEntity> existingGroupIdRecords = groupIdRecordDao.getAll();
         if (existingGroupIdRecords != null && !existingGroupIdRecords.isEmpty()) {
             for (GroupIdRecordEntity existing : existingGroupIdRecords) {
-                //Compare if it is a new element or if the element to compare dont have the same put code than me
-                if(newGroupIdRecord.getPutCode() == null || !newGroupIdRecord.getPutCode().equals(existing.getId())){
-                    if(newGroupIdRecord.getGroupId().equalsIgnoreCase(existing.getGroupId())) {
+                // Compare if it is a new element or if the element to compare
+                // dont have the same put code than me
+                if (newGroupIdRecord.getPutCode() == null || !newGroupIdRecord.getPutCode().equals(existing.getId())) {
+                    if (newGroupIdRecord.getGroupId().equalsIgnoreCase(existing.getGroupId())) {
                         throw new DuplicatedGroupIdRecordException();
                     }
-                }                                              
+                }
             }
         }
     }
